fix(http client): read body size from header.
Expermential to read number of bytes from `HTTP Content Length` to pre-allocate the number of bytes and bail early
if the length is bigger than the `max_request_body size`

Need to be benched with bigger requests.

+++ src/client/http/client.rs
@@ -4,13 +4,13 @@ use crate::types::jsonrpc::{self, JsonValue};
 use std::sync::atomic::{AtomicU64, Ordering};
 
 /// Default maximum request body size (10 MB).
-const DEFAULT_MAX_BODY_SIZE_TEN_MB: usize = 10 * 1024 * 1024;
+const DEFAULT_MAX_BODY_SIZE_TEN_MB: u32 = 10 * 1024 * 1024;
 
 /// HTTP configuration.
 #[derive(Copy, Clone)]
 pub struct HttpConfig {
 	/// Maximum request body size in bytes.
-	pub max_request_body_size: usize,
+	pub max_request_body_size: u32,
 }
 
 /// JSON-RPC HTTP Client that provides functionality to perform method calls and notifications.
